requirements.txt
aiohttp==3.0.2
async-timeout==2.0.0
attrs==17.4.0
beautifulsoup4==4.6.0
bs4==0.0.1
cffi==1.11.4
chardet==3.0.4
discord.py==1.0.0a1434+g355eb08
idna==2.6
idna-ssl==1.0.0
multidict==4.1.0
Pillow==5.0.0
pycparser==2.18
PyNaCl==1.2.1
six==1.11.0
websockets==4.0.1
yarl==1.1.1
